LPS-96136 Move com.liferay.portlet.internal tests to portlet-test
